Rename api_const to constants
