+++ test/cppsuite/CMakeLists.txt
@@ -2,7 +2,7 @@ project(cppsuite CXX)
 
 # Create an intermediate static lib.
 add_library(cppsuite_test_harness STATIC
-    src/common/api_const.cpp
+    src/common/constants.cpp
     src/common/logger.cpp
     src/common/random_generator.cpp
     src/common/thread_manager.cpp
