+++ test/cppsuite/src/component/component.cpp
@@ -28,7 +28,7 @@
 
 #include "component.h"
 
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 
 namespace test_harness {
