+++ test/cppsuite/src/component/runtime_monitor.cpp
@@ -31,7 +31,7 @@
 #include <fstream>
 
 #include "perf_plotter.h"
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 #include "src/storage/connection_manager.h"
 #include "statistics/cache_limit.h"
