+++ test/cppsuite/src/component/statistics/statistics.cpp
@@ -28,7 +28,7 @@
 
 #include "statistics.h"
 
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 #include "src/component/runtime_monitor.h"
