+++ test/cppsuite/src/component/workload_tracking.cpp
@@ -28,7 +28,7 @@
 
 #include "workload_tracking.h"
 
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 #include "src/storage/connection_manager.h"
