+++ test/cppsuite/src/main/database.cpp
@@ -29,7 +29,7 @@
 #include "database.h"
 
 #include "collection.h"
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/random_generator.h"
 #include "src/storage/connection_manager.h"
