+++ test/cppsuite/src/main/database_operation.cpp
@@ -29,7 +29,7 @@
 #include "database_operation.h"
 
 #include "src/common/thread_manager.h"
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 #include "src/common/random_generator.h"
 #include "src/component/workload_validation.h"
