+++ test/cppsuite/src/main/test.cpp
@@ -28,7 +28,7 @@
 
 #include "test.h"
 
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 #include "src/component/perf_plotter.h"
