+++ test/cppsuite/src/main/thread_context.cpp
@@ -28,7 +28,7 @@
 
 #include "thread_context.h"
 
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 #include "src/common/random_generator.h"
 #include "transaction.h"
