+++ test/cppsuite/src/main/transaction.cpp
@@ -28,7 +28,7 @@
 
 #include "transaction.h"
 
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/logger.h"
 #include "src/common/random_generator.h"
