+++ test/cppsuite/tests/search_near_03.cpp
@@ -26,7 +26,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "src/common/api_const.h"
+#include "src/common/constants.h"
 #include "src/common/random_generator.h"
 #include "src/main/test.h"
